+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Add DeploymentConfig sub-package for loading (and, optionally, decrypting) runtime environment configuration
+
 ### v1.1.0 (2019-07-09)
 
 * Allow asserting that the StoppedMockClock never slept